Move invite-events analytics into common/analytics and assert events fire once
