+++ frontend/src/metabase/analytics/index.ts
@@ -1,4 +1,3 @@
 export { createSnowplowTracker } from "./snowplow";
 export { trackPageView } from "./page-view";
 export { trackSchemaEvent, trackSimpleEvent } from "./event";
-export { trackInviteToViewOpened, trackUserInvited } from "./invite-events";
