+++ frontend/src/metabase/common/analytics/index.ts
@@ -1,1 +1,2 @@
 export { trackCustomHomepageDashboardEnabled } from "./homepage";
+export { trackInviteToViewOpened, trackUserInvited } from "./invite-events";
