+++ frontend/src/metabase/setup/actions.ts
@@ -1,11 +1,11 @@
 import { createAction } from "@reduxjs/toolkit";
 import { t } from "ttag";
 
-import { trackUserInvited } from "metabase/analytics";
 import { setupApi, userApi } from "metabase/api";
 import { loadLocalization } from "metabase/api/localization";
 import { isEmailAlreadyInUse } from "metabase/api/utils/errors";
 import { runRtkEndpoint } from "metabase/api/utils/run-rtk-endpoint";
+import { trackUserInvited } from "metabase/common/analytics";
 import { createDatabase } from "metabase/redux/databases";
 import {
   initializeSettings,
